// src/types.ts
export enum TAGS {
  BOLD = "BOLD",
  ITALIC = "ITALIC",
  H1 = "H1",
  H2 = "H2",
  H3 = "H3",
  H4 = "H4",
  H5 = "H5",
  H6 = "H6",
  UNDERLINE = "UNDERLINE",
  STRIKETHROUGH = "STRIKETHROUGH",
  HIGHLIGHT = "HIGHLIGHT",
  SUBSCRIPT = "SUBSCRIPT",
  SUPERSCRIPT = "SUPERSCRIPT",
  BLOCKQUOTE = "BLOCKQUOTE",
  LINK = "LINK",
  IMAGE = "IMAGE",
  FENCED_CODE_BLOCK = "FENCED_CODE_BLOCK",
  TABLE = "TABLE",
}